# Use vectorized row summaries