Replacing `rowwise()` summaries with vectorized ones when creating concepts results in a significant performance increase. For example the lookup for C0061 went from 275 s to 75 s in one dataset, and from over 600 s to 85 s in another dataset:

```r
health_records <- data.table::fread("/mnt/project/deep_phewas/health_records.txt.gz")
system.time(
  DeepPheWAS:::clinical_code_lookup("C0061", "all_cancer_codes_rated.csv.gz", health_records, "default")
)
```